requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bleach==6.2.0
blinker==1.4
certifi==2024.2.2
charset-normalizer==3.3.2
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.2.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.57.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
gyp==0.1
h5py==3.11.0
httplib2==0.20.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.6.0
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.4.2
kaggle==1.6.17
keras==3.3.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==8.10.0
moviepy==2.1.2
namex==0.0.8
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
oauthlib==3.2.0
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.3.6
proglog==0.1.11
protobuf==4.25.3
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.5.3
tqdm==4.67.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.29.1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==1.0.0
